+++ js/data.js
@@ -1,0 +1,32 @@
+import {rnd25} from './util.js';
+import {rnd200} from './util.js';
+import {rnd6} from './util.js';
+import {rndMessage} from './util.js';
+import {rndNames} from './util.js';
+import {getUserData} from './util.js';
+
+const arMessage = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+const arNames = [
+  'Иван',
+  'Александр',
+  'Степан',
+  'Валерий',
+  'Афанасий',
+  'Борис',
+];
+
+rnd25 ();
+rnd200 ();
+rnd6 ();
+rndMessage ();
+rndNames ();
+getUserData ();
+
