+++ js/util.js
@@ -1,0 +1,32 @@
+const rnd25 = () => Math.round(Math.random() * 24 + 1);
+const rnd200 = () => Math.round(Math.random() * 174 + 15);
+const rnd6 = () => Math.round(Math.random() * 5 + 1);
+
+
+const rndMessage = () => arMessage [rnd6() - 1];
+const rndNames = () => arNames [rnd6() - 1];
+
+const getUserData = () => {
+  const obj = {};
+  obj.id = rnd25();
+  obj.url = `photos/${rnd25()}.jpg`;
+  obj.description = 'Какое-то описание фотографии.';
+  obj.likes = rnd200();
+  const objComment = {};
+  objComment.id = rnd200(); //couldBeImproved
+  objComment.avatar = `img/avatar-${rnd6()}.svg`;
+  objComment.Message = rndMessage ();
+  objComment.Names = rndNames ();
+  obj.comments = [
+    objComment
+  ];
+  return obj;
+};
+
+export {rnd25};
+export {rnd200};
+export {rnd6};
+export {rndMessage};
+export {rndNames};
+export {getUserData};
+
